Progressbars for stats

+++ Legends/Legends/Models/LegendInfo.cs
@@ -6,10 +6,10 @@ namespace Legends.Models
 {
     public class LegendInfo
     {
-        public int Attack { get; set; }
-        public int Defense { get; set; }
-        public int Magic { get; set; }
-        public int Difficulty { get; set; }
+        public double Attack { get; set; }
+        public double Defense { get; set; }
+        public double Magic { get; set; }
+        public double Difficulty { get; set; }
         public string InfoToString()
         {
             return $"Attack: {Attack}\nDefense: {Defense}\nMagic: {Magic}\nDifficulty: {Difficulty}";
